Move the click behavior to the base class and add comment
